Separate vote listings rpc by protocol + fmt
