+++ rpc/src/services/protocol/proto_008/mod.rs
@@ -3,3 +3,4 @@
 
 mod helpers;
 pub(crate) mod rights_service;
+pub(crate) mod votes_service;
