+++ tezos/messages/src/protocol/proto_003/mod.rs
@@ -2,5 +2,6 @@
 // SPDX-License-Identifier: MIT
 pub mod constants;
 pub mod rights;
+pub mod votes;
 
 pub const PROTOCOL_HASH: &str = "PsddFKi32cMJ2qPjf43Qv5GDWLDPZb3T3bF6fLKiF5HtvHNU7aP";
