+++ tezos/messages/src/protocol/proto_004/mod.rs
@@ -2,5 +2,6 @@
 // SPDX-License-Identifier: MIT
 pub mod constants;
 pub mod rights;
+pub mod votes;
 
 pub const PROTOCOL_HASH: &str = "Pt24m4xiPbLDhVgVfABUjirbmda3yohdN82Sp9FeuAXJ4eV9otd";
