+++ tezos/messages/src/protocol/proto_005/mod.rs
@@ -2,5 +2,6 @@
 // SPDX-License-Identifier: MIT
 pub mod constants;
 pub mod rights;
+pub mod votes;
 
 pub const PROTOCOL_HASH: &str = "PsBABY5HQTSkA4297zNHfsZNKtxULfL18y95qb3m53QJiXGmrbU";
