+++ tezos/messages/src/protocol/proto_005_2/mod.rs
@@ -3,5 +3,6 @@
 pub mod constants;
 pub mod contract;
 pub mod rights;
+pub mod votes;
 
 pub const PROTOCOL_HASH: &str = "PsBabyM1eUXZseaJdmXFApDSBqj8YBfwELoxZHHW77EMcAbbwAS";
