+++ tezos/messages/src/protocol/proto_007/mod.rs
@@ -3,5 +3,6 @@
 pub mod constants;
 pub mod contract;
 pub mod rights;
+pub mod votes;
 
 pub const PROTOCOL_HASH: &str = "PsDELPH1Kxsxt8f9eWbxQeRxkjfbxoqM52jvs5Y5fBxWWh4ifpo";
