+++ tezos/messages/src/protocol/proto_008/mod.rs
@@ -3,5 +3,6 @@
 pub mod constants;
 pub mod contract;
 pub mod rights;
+pub mod votes;
 
 pub const PROTOCOL_HASH: &str = "PtEdoTezd3RHSC31mpxxo1npxFjoWWcFgQtxapi51Z8TLu6v6Uq";
